implemented many-to-one and one-to-one annotations into my entities classes
many-to-one annotaions are not working currently. the hibernate metadata generator is not correctly and I cannot use the code provided the hibernate 6 guide. need to learn about the metadata generator to get working. i have attempted to insert the write code into the pom.xml file to ensure the generator can be constructed